requirements.txt
Django==5.1
djangorestframework==3.15.1
channels==4.0.0
channels-redis==4.2.0
websockets==11.0
psycopg2-binary==2.9.9
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
daphne==4.0.0